requirements.txt
appdirs==1.4.3
Django==1.9.12
django-rest-framework-nested==0.0.1
djangorestframework==3.6.3
packaging==16.8
pyparsing==2.2.0
six==1.10.0
